SPI and Wire changes for arduino core api update. Include Hardware headers, change return type of Wire requestFrom from uint8_t to size_t

+++ libraries/Wire/src/Wire.cpp
@@ -66,7 +66,7 @@ uint8_t arduino::MbedI2C::endTransmission(bool stopBit) {
 	return 2;
 }
 
-uint8_t arduino::MbedI2C::requestFrom(uint8_t address, size_t len, bool stopBit) {
+size_t arduino::MbedI2C::requestFrom(uint8_t address, size_t len, bool stopBit) {
 	char buf[256];
 	int ret = master->read(address << 1, buf, len, !stopBit);
 	if (ret != 0) {
@@ -78,7 +78,7 @@ uint8_t arduino::MbedI2C::requestFrom(uint8_t address, size_t len, bool stopBit)
 	return len;
 }
 
-uint8_t arduino::MbedI2C::requestFrom(uint8_t address, size_t len) {
+size_t arduino::MbedI2C::requestFrom(uint8_t address, size_t len) {
 	return requestFrom(address, len, true);
 }
